removed redundant .typeValue() calls [3191]

+++ ucar/unidata/data/DataSourceImpl.java
@@ -968,7 +968,7 @@ public class DataSourceImpl extends SharableImpl implements DataSource,
         if (o == null) {
             return dflt;
         }
-        return Long.valueOf(o.toString()).longValue();
+        return Long.valueOf(o.toString());
     }
 
     /**
